@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.alert.alertapp'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.alert.alertapp"
         minSdk 24
-        targetSdk 33
+        targetSdk 34
         versionCode 1
         versionName "1.0"
 
@@ -29,11 +29,13 @@ android {
 }
 
 dependencies {
+    implementation "androidx.work:work-runtime:2.7.0"
     implementation files('libs/sqlite-jdbc-3.44.1.0.jar')
+    implementation "androidx.work:work-runtime:2.9.0"
     implementation 'it.xabaras.android:recyclerview-swipedecorator:1.2.2'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.8.0'
-    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'com.google.android.material:material:1.11.0'
+    implementation 'androidx.recyclerview:recyclerview:1.3.2'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
