@@ -1,5 +1,4 @@
 package com.alert.alertapp;
-
 import android.content.ContentProvider;
 import android.content.ContentUris;
 import android.content.ContentValues;
@@ -7,12 +6,9 @@ import android.content.Context;
 import android.content.UriMatcher;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
 import android.net.Uri;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-
 public class AlertProvider extends ContentProvider {
     public static SQLiteDatabase dB;
 
@@ -29,11 +25,11 @@ public class AlertProvider extends ContentProvider {
     }
     public Cursor getData() {
         SQLiteDatabase db = myDb.getWritableDatabase();
-        return db.rawQuery("SELECT * FROM " + "alerts1", null);
+        return db.rawQuery("SELECT * FROM " + "alarm", null);
     }
     public static ContentValues val;
     static final String CONTENT_AUTHORITY="com.alert.alertapp.alertprovider";
-    static final String PATH_ALERTS1="alerts1";
+    static final String PATH_ALERTS1="alarm";
     static final Uri BASE_CONTENT_URI=Uri.parse("content://"+CONTENT_AUTHORITY);
     static final Uri CONTENT_URI=Uri.withAppendedPath(BASE_CONTENT_URI,PATH_ALERTS1);
 
@@ -55,7 +51,7 @@ public class AlertProvider extends ContentProvider {
                 SQLiteDatabase db = myDb.getReadableDatabase();
                 String selection= "id"+"=?";
                 String[] selectionArgs={MainActivity.uId};
-                Cursor cursor = db.query("alerts1", strings, selection, selectionArgs, null, null, null);
+                Cursor cursor = db.query("alarm", strings, selection, selectionArgs, null, null, null);
                 if (cursor!=null && cursor.moveToFirst()){
                     int baslik=cursor.getColumnIndex("baslik");
                     String upbaslik=cursor.getString(baslik);
@@ -67,9 +63,15 @@ public class AlertProvider extends ContentProvider {
                     String utme=cursor.getString(time);
                     int leftTim=cursor.getColumnIndex("sure");
                     int leftTime=cursor.getInt(leftTim);
+                    int ringTon=cursor.getColumnIndex("ringtone");
+                    String ringTone=cursor.getString(ringTon);
+                    int rtm=cursor.getColumnIndex("realtime");
+                    //String realTime=cursor.getString(rtm);
                     cursor.close();
                     UptadeAlert.utext=upbaslik;UptadeAlert.usubText=upaciklama;
                     UptadeAlert.utime=utme;UptadeAlert.udate=udt;UptadeAlert.uLeftTime=leftTime;
+                    UptadeAlert.uRingtone=ringTone;UptadeAlert.hour=Integer.parseInt(utme.substring(0,2));
+                    UptadeAlert.minute=Integer.parseInt(utme.substring(3,5));//UptadeAlert.uRealTime=realTime;
                 }
                 return cursor;
             default:
@@ -92,13 +94,16 @@ public class AlertProvider extends ContentProvider {
                 String subText=AddAlert.subText;
                 String d=AddAlert.date;
                 String t=AddAlert.time;
+                String ringTone=AddAlert.ringTone;
+                String realTime=AddAlert.realDateTime;
                 int leftTime=AddAlert.leftTime;
-                Alerts alert=new Alerts(text,subText,d,t,leftTime);
+                Alerts alert=new Alerts(text,subText,d,t,leftTime,ringTone);
                 val=new ContentValues();
                 val.put("baslik",alert.getBaslik());
                 val.put("aciklama",alert.getAciklama());val.put("tarih",alert.getTarih());
                 val.put("saat",alert.getSaat());val.put("sure",alert.getLeftTime());
-                long ID=dB.insert("alerts1",null,val);
+                val.put("ringtone",alert.getRingTone());val.put("realtime",realTime);
+                long ID=dB.insert("alarm",null,val);
                 alert.setId(ID);
                 dB.close();;
                 MainActivity.source.openDb();
@@ -116,7 +121,7 @@ public class AlertProvider extends ContentProvider {
         int rowDelete=0;
         switch (matcher.match(uri)){
             case 1:
-                rowDelete=dB.delete("alerts1",s,strings);
+                rowDelete=dB.delete("alarm",s,strings);
                 break;
             default:
                 throw new IllegalArgumentException("Bilinmeyen Uri:"+uri);
@@ -124,17 +129,24 @@ public class AlertProvider extends ContentProvider {
         }
         return rowDelete;
     }
-
     @Override
     public int update(@NonNull Uri uri, @Nullable ContentValues contentValues, @Nullable String s, @Nullable String[] strings) {
         SQLiteDatabase dB=myDb.getWritableDatabase();
         ContentValues values=new ContentValues();
         values.put("baslik",UptadeAlert.utext);values.put("aciklama",UptadeAlert.usubText);
         values.put("tarih",UptadeAlert.udate);values.put("saat",UptadeAlert.utime);
-        values.put("sure",UptadeAlert.uLeftTime);
-        dB.update("alerts1", values, "id=" +MainActivity.uId, null);
+        values.put("sure",UptadeAlert.uLeftTime);values.put("ringtone",UptadeAlert.uRingtone);
+        values.put("realtime",UptadeAlert.uRealTime);
+        dB.update("alarm", values, "id=" +MainActivity.uId, null);
         dB.close();
         return 0;
     }
-
+    public static boolean isDatabaseEmpty() {
+        // Hataya neden olabilecek bir durumu kontrol et
+        if (dB == null) {
+            return true;
+        }
+        else
+            return false;
+    }
 }
