@@ -1,5 +1,4 @@
 package com.alert.alertapp;
-
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
@@ -7,6 +6,6 @@ import android.content.Intent;
 public class AlertReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
-        
     }
+
 }
